Move blockchain tests to new files
